.......... [ZBXNEXT-686] improved performance of the backup/restore operations with PostgreSQL

+++ ui/tests/include/helpers/CDBHelper.php
@@ -241,15 +241,18 @@ class CDBHelper {
 
 		$suffix = '_tmp'.count(self::$backups);
 
-		foreach ($tables as $table) {
-			DBexecute('DROP TABLE IF EXISTS '.$table.$suffix);
+		if ($DB['TYPE'] === ZBX_DB_POSTGRESQL) {
 
-			switch ($DB['TYPE']) {
-				case ZBX_DB_MYSQL:
-					DBexecute('CREATE TABLE '.$table.$suffix.' AS SELECT * FROM '.$table);
-					break;
-				default:
-					DBexecute('SELECT * INTO TABLE '.$table.$suffix.' FROM '.$table);
+			if ($DB['PASSWORD'] !== '') {
+				putenv('PGPASSWORD='.$DB['PASSWORD']);
+			}
+			$server = $DB['SERVER'] !== '' ? ' -h'.$DB['SERVER'] : '';
+			exec('pg_dump'.$server.' -U'.$DB['USER'].' -Fd -j5 -t'.implode(' -t', $tables).' '.$DB['DATABASE'].' -f '.PHPUNIT_COMPONENT_DIR.$DB['DATABASE'].$suffix.'.dump');
+		}
+		else {
+			foreach ($tables as $table) {
+				DBexecute('DROP TABLE IF EXISTS '.$table.$suffix);
+				DBexecute('CREATE TABLE '.$table.$suffix.' AS SELECT * FROM '.$table);
 			}
 		}
 	}
@@ -268,22 +271,29 @@ class CDBHelper {
 		$suffix = '_tmp'.count(self::$backups);
 		$tables = array_pop(self::$backups);
 
-		if ($DB['TYPE'] == ZBX_DB_MYSQL) {
+		if ($DB['TYPE'] === ZBX_DB_POSTGRESQL) {
+
+			if ($DB['PASSWORD'] !== '') {
+				putenv('PGPASSWORD='.$DB['PASSWORD']);
+			}
+			$server = $DB['SERVER'] !== '' ? ' -h'.$DB['SERVER'] : '';
+			exec('pg_restore'.$server.' -U'.$DB['USER'].' -Fd -j5 --clean -d '.$DB['DATABASE'].' '.PHPUNIT_COMPONENT_DIR.$DB['DATABASE'].$suffix.'.dump');
+			exec('rm -rf '.PHPUNIT_COMPONENT_DIR.$DB['DATABASE'].$suffix.'.dump');
+		}
+		else {
 			$result = DBselect('SELECT @@unique_checks,@@foreign_key_checks');
 			$row = DBfetch($result);
 			DBexecute('SET unique_checks=0,foreign_key_checks=0');
-		}
 
-		foreach (array_reverse($tables) as $table) {
-			DBexecute('DELETE FROM '.$table);
-		}
+			foreach (array_reverse($tables) as $table) {
+				DBexecute('DELETE FROM '.$table);
+			}
 
-		foreach ($tables as $table) {
-			DBexecute('INSERT INTO '.$table.' SELECT * FROM '.$table.$suffix);
-			DBexecute('DROP TABLE '.$table.$suffix);
-		}
+			foreach ($tables as $table) {
+				DBexecute('INSERT INTO '.$table.' SELECT * FROM '.$table.$suffix);
+				DBexecute('DROP TABLE '.$table.$suffix);
+			}
 
-		if ($DB['TYPE'] == ZBX_DB_MYSQL) {
 			DBexecute('SET foreign_key_checks='.$row['@@foreign_key_checks'].',unique_checks='.$row['@@unique_checks']);
 		}
 	}
